requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sqlite3
SpeechRecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1
comtypes==1.1.14